requirements.txt
Django==3.2.13
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter==2.4.0
mysqlclient==2.2.7
psycopg2==2.9.10
tqdm== 4.67.1
